chatbot/chatbot_requirements.txt
python-dotenv

pydantic
pydantic-settings

pinecone

langchain
langchain-pinecone
langchain-openai
langchain-huggingface

langserve
langgraph

fastapi
uvicorn

sse_starlette
